Implement paginated repository listing and caching system
Infinite Scroll system with Pagging 3 was implemented.
Cache system for API listing was implemented.
Use Cases were created for the application.
The correct display of the list of items on the screen was made.
Mocks and false data were removed in the application, passing real data from the API.
Removed and replaced application magic numbers.